+++ tests/corpus/io/test_musan.py
@@ -25,6 +25,8 @@ class TestMusanDownloader:
         downloader = io.MusanDownloader()
 
         with requests_mock.Mocker() as mock:
+            # Return any size (doesn't matter, only for prints)
+            mock.head(musan.DOWNLOAD_URL, headers={'Content-Length': '100'})
             mock.get(musan.DOWNLOAD_URL, content=tar_data)
 
             downloader.download(target_folder)